requirements.txt
google-api-python-client==2.102.0
google-auth-httplib2==0.1.1
google-auth==2.23.3
Pillow==9.0.1
streamlit==1.23.1

